requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
beartype==0.18.5
et-xmlfile==1.1.0
fast_bitrix24==1.7.1
frozenlist==1.4.1
icontract==2.6.6
idna==3.7
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
yarl==1.9.4
